Proposal data adaption workaround
Fix #125, this applies to the last hour of the proposal. If overrides by
default to the fist hour of the proposal, both pointing to 00:00 but to
diff days.

The workaround moves to the end the last hour of the proposal.

Concretely, if current hour exits and have a value, set current hour to 25.

That's not the ending solution, but provides a quick fix.

The final solution will be something similar to group measures in a list
of days, and paint iterating the sorted list of days.